[build] Add support for tests with multiple go packages
Introduces new gopackages = ["pkg1", "pkg2"] feature that can be
used as an alternative to gopackage = "pkg" for tests.

Motivation:
FIDL has a large number of tests and it can be difficult to sort out which command to run.
We want to simplify things and combine some of the tests. For instance, the fidlgen c++ and
fidlgen c++ ir tests live in different packages but you very rarely would need to run them
separately.

FIDL-766 #comment

Tests: manually -- modified gn files to use gopackages rather than
packages and also tested the case where both are specified (which
errors).

Change-Id: I58ba62db8f1cc2f6f4d50afa7512aae9e2beecfb