[debugger] Fix infinite loops on message loop.
Because signals have to cleared, the message loop would forever loop
over the task signal.

This also handles the case where ZX_SOCKET_WRITABLE would also be
triggered inifnitively and now go ask for the signal only when
necessary.

On top it would show the debugger using 99.9% of the cpu.

DX-1824 #done

TEST=Manual

Change-Id: Ic7b61c381bc04c2b8d4dfae77837b80fb704c948