[terminal] launch and connect to shell.
The Pty class manages the connection to the shell.
It launches the shell and provides access to a file
descriptor which can be used to communicate with the
shell.

The current architecture of the view controller
requires that we make an assumption about the size
of the window so we can launch the shell and set up
our io loop. This will be refactored in a subsequent
CL which will introduce an event loop. The terminal
redraws the entire screen for each read which is not
performant. This will also be addressed in a later CL.

TEST=fx run-test terminal_tests
- Tests that the server is spawned and can be closed
- Tests that the pty can read/write
- Tests that vim and ls can be launched
- Tests that the window can be resized

Change-Id: Ia1b74ac624154feb2de783ab669fc2814a4da179