[debugger] Add build dir options.
Hooks up the build directory option to the code that finds source files.

Adds the DWARF "compilation dir" to the FileLine class and makes the
"file" relative to this compilation directory rather than resolving it
when reading names from the symbols. This means that we have all
information possible available at the point of needing to open the file.

Adds a fair bit of fallback logic to search for files in different
places since the symbol paths aren't strictly defined.

Adds a command line option to add to the build directory list. Currently
this falls back on the hardcoded one which we should remove eventually.
This command line option is not yet passed by any scripts.

Adds a ScopedTempDir helper for making the test easier to write.

Change-Id: I34d777ab36c5b814ae982338a18936ba70d85ace