[debugger] Hold an error or a value.
This class can be used to replace an Err and a value type.

I was originally trying to avoid this type of class, even though it is
relatively common, because I felt like it can add a layer of obscurity
to everything.

But the number of cases where we have to pass an (Err, Value) pair and
manually default construct the "other" case is very large and becoming
bothersome. I think this will help clean up the callbacks in the
expression system.

Change-Id: I540a9c1053a8fd16e38dfab849c0d5ec46c9b0da