[netstack] Fix crash on UDP connect to unspecified
When connect() is called on a UDP socket to the unspecified address, the
socket is not given a remote address, causing one of our logging
statements to return an error, which we panic on. This adds a check if
the address is the unspecified address before attempting to log the
local and remote addresses.

Also fix handling of AF_UNSPEC while I'm here.

Change-Id: I54c05d1d44c31773bbb440827c323e87461cf381