[audio_core] API to provide AudioCore a channel
Devices that are not backed by the traditional backplane may
need their implementations to be services, not drivers. This
CL enables a service to provide a stream channel to the
AudioCore service, which should then attempt to load it just
like any other driver-backed stream channel.

This CL also includes a new test binary audio_device_service_tests
that targets this new API in the AudioDeviceEnumerator protocol.

MULTIPLY: `[{
    "target": "//src/media/audio/audio_core/test:audio_device_service_tests",
    "total_runs": 100
}]`

Test: build, CQ (including new test bin)

MTWN-196 #done

Change-Id: Ice444d5d4265ee4e4eaf67ef16fa701ea0a68f08