[perfcompare] Test CIs produced from data with variation at different levels
The format for perf test results accepted by perfcompare as input has
got gradually more complicated.  It started off as 1-level (for
multiple iterations within a process) but has changed to be 3-level
(for multiple boots, multiple process launches, and multiple
iterations within a process).

Add a test that generates test data with variation at each of those 3
levels, and check the CIs produced from that test data.

This is in preparation for changing how the CIs are calculated.

Issue: PT-202
Test: "python perfcompare_test.py" (run by bots on CQ)
Change-Id: I2382ec2a592a960b996cad10d15ea8c9009c99dd