[kernel][port][hypervisor] Return BAD_HANDLE when queuing to a closed port
Prior to this change, queuing a packet to a port concurrent with
closing the last handle to the port could result in either
ZX_ERR_BAD_HANDLE or ZX_ERR_BAD_STATE depending on the timing (before
or after the handle table lookup).

After this change, we'll return ZX_ERR_BAD_HANDLE in both cases.

In the hypervisor, when queuing to a closed port via Trap::Queue,
continue returning ZX_ERR_BAD_STATE as that status is a better match
than ZX_ERR_BAD_HANDLE as there is no handle in the syscall that
initiated the Trap::Queue (e.g. zx_vcpu_resume).  No change in
behavior for hypervisor.

Bug: ZX-4766 #done
Test: added tests to /boot/test/core/port-test and /boot/test/sys/hypervisor-test
Change-Id: I90dc8cd7c3ddbf3dbc45f542889d9069415e59a4