[rust_crates] Update nom to 5.0
Also allow it to be used by host targets, so that it can be used from
the driver bind compiler (//tools/bindc). And enable the rustexp_macros
feature (causes a dependency on regex and lazy_static), which will also
be used in the bind compiler tool.

Ran:
$ fx update-rustc-third-party
$ fx build
$ fx update-rustc-crate-map --output third_party/rust_crates/crate_map.json

Updated WLAN uses of nom. Mainly fixes to the error handling and a few
small changes due to differences in how many0 interacts with parsers
that may be Incomplete (i.e. streaming parsers).

Ran the unit tests on each of the WLAN packages I changed.

Change-Id: Iaffa5fcef6ef162d877a4a4cd93e1a90ad5c9ee4