[ledger] Retrack identifiers upon storage reset
Object identifiers are bound to a specific page. Some of our tests use a
first dummy page to compute object identifiers and populate some fake
remote data sources, and then instantiate a second page and use the same
identifiers with it. This is now unsound, and makes test fail when
tracking is enabled.

This commit recomputes fresh identifiers tracked by the new page based
on the data from the old identifiers.

Bug: LE-702
Test: fx run-test ledger_tests -t ledger_unittests
Change-Id: I5b8736f42b970de53a6cf27ebdfdba2dedfb381e