[zircon][crashsvc] Route crashsvc to ExceptionBroker
This changes the service crashsvc uses when it receives an unhandled
exception. Before it was sent directly to CrashpadAgent via
fuchsia.crash.Analyzer.

Now it is send to ExceptionBroker via fuchsia.exception.Handler, which
in turn passes it to CrashpadAgent using the Analyzer API. This
effectivelly "injects" ExceptionBroker in the middle.

This will permit to be able to route exceptions to other handlers
besides CrashpadAgent (eg. a debugger).

DX-1784 #done
DX-1757 #comment Exception broker will have the ability to route to a
possible debugger.

TEST=
Updated crashsvc tests to this new API:
- fx shell -- runtests -t crashsvc-test

ExceptionBroker has both unit and integration tests:
- fx run-test exception_broker_tests

Tested manually (seeing the log output):
- fx shell -- crasher

Change-Id: I48c4c1ad2f5ed2ca1eaccd53a16eeacc46ecd109