[a11y] Rename SemanticActionListener -> SemanticListener and add a method to set the update mode.
Following up on rename requested by API review.

Renaming a protocol is a breaking change, so this is one of those "gets worse before it gets better" sorts of things. This renames the SemanticListener and adds a new version of RegisterView that uses it. The OnModeChange method has been added to the new SemanticListener so that we can tell the client whether or not to send updates. SemanticActionListener will be removed once all references to it are gone.

Testing: * Build: (Passed)
  * Unit Test: fx run-test a11y_tests
Change-Id: I5b8f9bf366f214921320c895a7e965db3bcc247b