[symbolize] Output human readable contextual info
This change makes the symbolizer output a human readable form of the
contextual information in a compacted form. Prior to this change the
symbolizer would just remove all references to decrease noise. This
made debugging certian things require the old output. The format
used here is an attempt to mimic the existing output format but with
one extra field for the module ID. This is in case we have to output
the mmap information later. See ExampleDemux for how this is handled

Basically if we have soemthing like the following

```
{{{module:0:...:libc.so}}}
some text
{{{mmap:0x12345000:0x1000:load:0:...}}}
some other text
```

The expected output is

```
[[[ELF module #0x0 "libc.so" BuildID=...]]]
some text
[[[ELF seg #0x0 0x12345000]]]
```

This allows module and mmap statements to be output out of order and the
information is still avliable to a human.

Change-Id: Ifde7513b993d49ef277b0bb7a4ce63b1f8628d3f