[codec] VP9 decode conformance testing
With some fixups to the HW VP9 decoder included in this CL, many of the
conformance test streams pass, but not all do yet.

This CL also changes the input path of use_video_decoder to permit
feeding from a file or from http.  Both are now fully streaming.
Streaming via http is required to deal with the larger test streams
without hitting target storage resource limits, when testing multiple
streams in an e2e style test pass.

For now, we avoid increasing the amount of contig RAM reserved up front,
but this means that some of the larger VP9 decode conformance streams
won't work yet.  There's some frame count plumbing that'll help some
here, to get the number of needed frames to be < 16.

Change-Id: I586d30de35c35e708298dba2a236daafdb301ef5