[debugger] Debug probe rubustness improvements.
Recently Rust files stopped being able to be symbolized.

The problem is that the .build-id file with no ".debug" extension has no
debug information but does have a section names ".debug_gdb_scripts"
which triggers the elflib debug info probe and we use this file.

This patch fixes this bug on two ways:

- Looks in the "<build-id>.debug" file first. This file's definition
  means it's more likely the correct file if it contains program
  information (which we prefer).

- Makes the elflib debug info probe specifically match only
  ".debug_info" sections. The previous code was trying to be a general
  problem for *any* debug information. But there are many different
  kinds that could be stripped or not (as the Rust toolchain is doing).
  To know when a file contains what you need, you need to query that
  specific thing. A comment explaining this is added to the probe
  function.

Change-Id: Ib925ba4d022fb8e2517406d9443329599ecb0a5e