[wlan testing] wlan-smoke-test use http_service_util
Move call from the locally implemented getch and discard method to the
connectivity testing crate service.  Additionally adds a check for each
wlan interface and a unit test verifying the returned result from the
helper crate.

Note: Since we are still unable to verify that the data transfer is
using a specific interface, this does not guarantee the wlan interface
tested is carrying the data.  Further APIs are needed to perform this
check.  The current state is valuable given the connectivity lab setup
and on-device testing instructions.

Bug: WLAN-1024
Test: fx run-test connectivity-testing-tests
Test: fx run-test wlan-smoke-test-tests
Test: manually verified wlan-smoke-test still works
Change-Id: I2d37414a6980706cb6c6ff59201889e967646222