[build] Enable debugdata in sandbox when using profile variant
We need to allow the access to fuchsia.debugdata.DebugData service
from the sandbox when using the profile variant to allow publishing
the profile data.

This doesn't currently support the case where selector is being used
to apply the profile variant only to a single target, but it enables
the collection of profile information on bots where we always apply
variant for the entire build.

Bug: CF-698
Change-Id: I1e64e8342d196678859cf7b910fcc70a96bd733e