[audio_core] Enable audio CHECK for NDEBUG builds
This CL is designed to be rapidly reverted when we see the first
occurence of an elusive condition that we have only experienced
in CI.

Specifically, this CL converts a DCHECK into a CHECK, just in
case the problem is more likely to be hit on NDEBUG builds.

Test: assert-only change that is already present in DEBUG builds,
no other change.

MTWN-336 #comment Changed DCHECK to CHECK

Change-Id: I17060c6f822f59003e6ef93601deca075bb29176