[fidlc] 32 to 64 bits Ordinals: Step #2
Shifting gen_ord << 32, and expose as uint64 in JSON IR. All bindings
further shift this value << 32 such that the resulting ordinal bindings
see is 0, and will be ignored.

Re-generated the goldens to ensure this has no impact

	fx exec garnet/go/src/fidl/compiler/backend/typestest/regen.sh

Only the JSON IR is changed (as expected).

Moving fidlcat forward to Step #3, where the bindings (in this case
fidlcat) read the gen ordinal without shifting.

FIDL-524 #comment

Test: json_generator_tests.cc
Change-Id: I594dc0e78c0654344e21cf1b52bf4514175947c4