[Brightness] New Brightness Manager
This manager controls the manual and auto-brightness
for the display backlight.

This is an initial version to cover the high priority case of having
auto-brightness working.

Still to do: Hanging get protocol.
Still to do: More tests.
Still to do: Change coded driver limits when drivers change.

Design: go/fuchsia-brightness-design
Test: Wave hand in front of device, the screen changes appropriately.
Test: Unit tests have been started.
Test: Test client written that switches between auto/manual and sets manual values.
Change-Id: I2561afe02b1877075c5ba7dcd0c05063b6e46683