[debroot] Use single sysroot for all architectures
Rather than producing separate sysroots for different architectures,
we produce a single sysroot that contains all architectures and rely
on multiarch support in the compiler to pick up the right files for
each target.

Additionally, this change also includes libxi package which is needed by
glfw3.3, removes libgcc which is no longer needed now that compiler-rt
provides crtbegin.o/crtend, and switches over to the snapshot mirror as
the source for packages.

Bug: TC-562
Change-Id: I5eee2c47424e1720878e9b43b47428326ef9e9c1