[i2c-hid] Convert to c++ and add tests
ZX-4730 #done

Convert to c++. The functionality should be
identical. There is now an integration
test which checks that the driver comes
up and is initialized properly.

This patch is re-adding the similar patch that was
reverted (I7de9ebe81d476a4c4e92f5ce8dd06d6a43d49e0e).

This patch now fixes a bug where an error in the worker
thread prevented the driver from shutting down.
This is tested with 'HidTestReportDescFailureLifetimeTest'.

TEST: runtests -t i2c-hid-test-test
Change-Id: I27abbc0a88731a565bdb3b638328dfbc0a009877