[syscalls][fidl] Preliminary expression of syscalls in fidl
This is a first attempt at expressing syscalls in .fidl. There are many
places that will need to change. Constraints, rights, structs, various
formulations of input/output parameters are incomplete or non-existent.

i.e. these will require much more further iteration, but have just
enough structure that a subsequent change to land the backend will be
able to generate one of the outputs that abigen currently generates.

This is not used by the build and no one needs to update any of these
files. This is noted explicitly in the README of this directory and at
the top of every file as well. Additionally, they're named with the
.test.fidl suffix rather than just .fidl to 1) help indicate that
they're in-progress and not in use; and 2) not require API Review until
they're closer to actually being used.

ZX-3882 #comment [syscalls][fidl] Preliminary expression of syscalls in fidl

Change-Id: I737e8a9e031e69bba805da1b9a1086077a76c78f