[debugger] Handle more cases for the ld debug addr.
Adds support for the case where the dynamic loader is already
initialized at process startup. This means that the property that
indicates the debug address is already valid and should not be
overwritten with the value that signals ld.so should break on
initialization.

A more complete description of the operation of this, both the old case
and this new edge case is added to debugged_process.cc

TEST=There should be no behavior change in the normal case. The new
behavior is only visible with some experimental code that hasn't landed
that clones the ld.so state before the process begins execution. This
was tested manually in this environment.

Change-Id: Ib93dd4ffb4e750281c891f54bcb18fa36a5dc439