[modular][basemgr_launcher] Allow specific arguments
This allows basemgr_launcher to accept a base_shell, a session_shell, and/or
a single session_agent argument for benchmark testing via shell scripts.

This also changes basemgr to hold onto a copy of the Modular
configuration and host it for sessionmgr. Previously, basemgr was cloning
the directory it was reading configs from and passing that along to sessionmgr.
This doesn't work when basemgr_launcher starts basemgr and dies before
sessionmgr is initialized because the config directoy won't exist.

Both of these changes are tested in basemgr_launcher_tests.
TEST=fx run-test basemgr_launcher_tests
MF-491 #done

Change-Id: Id23b323285f1769a88d3ea94354bf8d6e8ca097f