[make-integration-patch] Support non-system bash interpreter
Switch from the absolute '#!/bin/bash' to '#!/usr/bin/env bash' so
that we can use a non-default bash. This is needed on OSX as OSX's
default bash is quite old, and SIP prevents overwriting the system
bash interpreter.

Switching to 'env bash' resolves this problem by portably selecting
the user's own bash.

Change-Id: Ib26cb4ebb6ec8b1ff153b95b96e5e16373fb883f