[syscalls] don't copy handles on error
Currently, if a syscall fills a user_out_handle* parameter but then
fails, the filled handle will still get copied to userspace.

This is most commonly possible when creating handle pairs, such as with
zx_channel_create(). If the first handle is successfully copied out, but
then the second fails, the syscall will wind up still copying the first
handle out to userspace, which will likely leak since the caller won't
be expecting a valid handle on failure.

This CL changes the syscall wrapper to only copy handles out on syscall
success. This means that syscalls won't be able to return handles and a
non-OK status, which is a rule our current API appears to follow but
wasn't explicitly enforced until now.

Also updates the abigen tests to C++/zxtest as long as we're here.

New generated syscall code looks like this:

  syscall_result wrapper_event_create(uint32_t options, zx_handle_t* out, uint64_t pc) {
      return do_syscall(ZX_SYS_event_create, pc, &VDso::ValidSyscallPC::event_create, [&](ProcessDispatcher* current_process) -> uint64_t {
          user_out_handle out_handle_out;
          auto result = sys_event_create(options, &out_handle_out);
          if (result != ZX_OK)
              return result;
          if (out_handle_out.begin_copyout(current_process, make_user_out_ptr(out)))
              return ZX_ERR_INVALID_ARGS;
          out_handle_out.finish_copyout(current_process);
          return result;
      });
  }

ZX-3422

Change-Id: Iee43f580c074dc4b5291e2b0c7bae4d185dd8ee5