[audio_core] Adust logging for output skipping
Audio_core contains a set of timing-related logging messages,
to provide insight about underruns that may not be audible.
A particular message (indicating that an audio output stream
skipped some number of frames) has two possible causes, one
of which is entirely normal: on playback start, the lack of
a previously-played frame means we have no expectation about
when this new frame should have arrived. This "discontinuity"
is expected.

This CL clarifies the wording of a few of these messages, and
specifically reduces the logging level of this message to
TRACE (not displayed by default on DEBUG or NDEBUG builds).

Test: logging-only, no product behavior change

Change-Id: Ifd5a5b0c1b7e71eaebe3cb850d1c5b43c69236e6