[ledger] Do not require PageStorage to outlive Commit
Refactor CommitImpl and LiveCommitTracker into CommitFactory. This
factory handles the creation of new commits, and keeps track of
currently live ones. This allows us to check whether a commit is "alive"
(generated by a non-destructed PageStorageImpl object) by use of a WeakPtr
that is not exposed outside of the class.

Change-Id: Id14b99dbdd56dcf3d303c02ee895263532075dca