[wlan][mlme] Send WLAN frames from Rust
BUG: WLAN-1267 #done

`ClientStation` is an early start of what will be
a Rust written Client implementation.
For now, the STA contains references to the underlying
device for sending frames, a sequence manager, a
buffer provider, as well as BSSID and iface's MAC address.
This allows to simplify MLME's C-ABI significantly.
In addition, the SequenceManager is fully owned and managed
in Rust.
This allows to drastically simplify the wlan-mlme-c.

Test:
wlan-mlme-c is tested by existing C++ MLME tests.
Added tests for wlan-mlme and wlan-common
hw-simulator's integration tests verify entire stack changes.
Manual verification with ath10k

Change-Id: Ib5afc1a6577f35e2fed6897a8bafaad3aeaaec68