[a11y] Re-add size limits in A11y Fidl.
Size limits in A11y Fidl were added in
https://fuchsia-review.googlesource.com/c/fuchsia/+/306455

This change broke the builder because of a test in flutter because of
which it was rollbacked in
https://fuchsia-review.googlesource.com/c/fuchsia/+/307862

Current cl is re-adding size limits in A11y fidls.
This change will be submitted once flutter test is modified to handle
these limits.

Testing:
  * Build: Passed
  * Unit Test:
  	* fx run-test a11y_tests (Passed)
	* fx run-test flutter_runner_tests (Passed)

Change-Id: Iade21edea909320805f9c529cbf21fc418222bd6