[usb-virtual-bus] switch from serial-impl to serial
Prior to this change, the InitFtdi method would attempt to bind directly
to the ftdi serial-impl driver as a file. This caused issues with other
drivers that attempt to bind to the main serial driver, and caused
conflicts with reads and writes for usb-virtual-bus-ftdi-test.

This change addresses the issue by changing InitFtdi to bind instead to
the serial driver, which guarantees only one actively bound client. This
resolves the issue with usb-virtual-bus-ftdi-test.

Change-Id: I8825ce8a2da7eedbcd71c940beab86ec02002841