[runtests] Truncate test process name from the left
Process names in zircon can only be 32 characters. Test processes get
the pkgfs path as their name, and the interesting bits of that path are
to the right, so instead of just letting the name get chopped off at 32
characters, we should chop it ourselves, and chop the left side of the
path so the parts with the name of the actual test stay in the process
name.

Change-Id: I97ae9ae7d4adf503e63e973f38c0874b54a6333b