[activity-service] Implement activity state machine.
ActivityStateMachine is a simple state machine which currently just
drives between IDLE and ACTIVE based on user inputs and timeouts.
Additional states will be added later.

StateMachineDriver is a driver which manages an ActivityStateMachine and
drives it based on user activity (or inactivity in the case of
timeouts). StateMachineDriver wraps around an async::Loop to ensure
sequential transitions and thread-safety in the presence of asynchronous
timers or events from many simultaneous clients.

Tested: Unit tests for StateMachineDriver.

Change-Id: Ic0e6fd47fbec97d30f40568b0b9f1905f27aaf07