[zircon][job][oom] Avoid kernel panic when oom kills empty job
If ZX_PROP_JOB_KILL_ON_OOM was set on an empty job, then when the OOM
killer thread called Kill(), it would transition immediately to DEAD. In
that case, it would attempt to remove it from all_jobs_list_. However
ForEachJob() is already holding that to iterate that list, so the mutex
would panic.

To resolve this, remove ForEachJob(), replacing its two uses with
something different: In diagnostics, use JobEnumerator. For KILL_ON_OOM,
add a new method KillJobWithKillOnOOM() which attempts to kill a child
of the given job that has get_kill_on_oom() set. This method doesn't use
JobEnumerator to avoid allocating during traversal (other than stack
space, which is limited to the maximum height of the job tree).

Additionally, add unit testing for this behavior (as `k ut job`).

ZX-4284 #done

Change-Id: I7da8f28eb9480abd389efb08e0c4fee909b4eeb4