[oom] Stow a crashlog on OOM
Previous behaviour on OOM was to simply reboot after allowing userspace
a short attempt at cleanup. Now, stow a crashlog (of a similar style to
kernel panics) before rebooting.

Post-reboot, this will cause the crash reporter (kernel_crash_checker)
to trigger a collection of relevant data and upload to the crash server.

In a follow up, we will also be able to further discriminate between
plain panics and ooms so that they can be filtered as such on the crash
server by changing the ptype field.

ZX-4409 #comment [oom] Stow a crashlog on OOM

Change-Id: Ie634db8634885eeb0c5820e542d58087906aa03d