[hwreg] Don't rely on compiler optimizations to ensure register state is small
Previously, we were relying on the compiler optimizing out the unused
1-byte Field members being created, which can only happen when an
instance of the register does not escape from a function.  If it did, it
could have a rather large size (especially if printers are enabled).

This improves the logic to give us smaller state if the compiler
supports [[no_unique_address]].  While this is a C++20 feature, Clang and
GCC support this for C++11 and later.

Change-Id: I87aab2d4743775cd238d4c041d3ec55f534fd86e