[zbi] Add option to ignore missing manifest files.
This option makes it possible to ignore missing files in a manifest.
This can be used when generating ZBI files in an environment where files
are created conditionally and are not guaranteed to exist.

This change also uses clang-format to do some autoformatting. As a
result, some headers have been reorganized.

ZX-4285 #done

TEST: ./zbi -o /tmp/test.zbi -i manifest_with_some_files_missing

Change-Id: I6430aa0bfbcea5bdd9d41b2df842335a48fceb92