[sysmem] Use ZX_VMO_CHILD_SLICE for safe recycle
Use ZX_VMO_CHILD_SLICE and ZX_VMO_ZERO_CHILDREN to detect when it's
actually safe to reclaim space.

Also zero and cache flush allocated VMOs before handing them out.  This
becomes more relevant now that there's more rigor in when we recycle.

This unfortunately won't work yet for VDEC - see ZX-4809 for more on
that.  For this reason the previous temporary mechanism is preserved by
this CL for now.

Change-Id: Ic6f3c50b2c523e13b4a113109526a26ce124b3af