[debugger] "attach" and "run" create new targets
With the more recent changes to attach filters where we automatically
create new targets more often, the requirement that the selected process
not be running doesn't make sense.

This allows "run" and "attach" to create a new target if the current one
is running. If the user manually specifies a process number, that one
will be forced and will still throw an error if it's running.

TEST=manually run and attach in different configurations.

Change-Id: I19c2e5e44fb0227e9debca82dfb0da9ea221bfe7