[kernel][vm] Allow non VmObjectPaged children
Assuming that children were of type VmObjectPaged made sense previously
as the only kind of child was copy-on-write. Now that children can also
be slices it is meaningful to allow different child types.

In this change it is still assumed that VmObjectPaged will only have
VmObjectPaged children, and so most of the non type signature changes
involve adding assertions and comments that this is in fact the case.

Change-Id: I725fba3633e8d645b77a478f187ef4335eb66b32