[a11y] (3/7)Hit Test Support in SemanticManager
This is the CL-3 in the long chain(Total 7) of CLs for adding Explore action
support on Screen Reader. This cl depends on
https://fuchsia-review.googlesource.com/c/fuchsia/+/299855 for review.

This cl adds changes to semantics_manager to support HitTest().
HitTest() is implemented by Semantic Tree provider(Like Flutter and
Chrome).

Given a Koid and Local Point, semantics manager will find the
appropriate View and call the HitTest API.

This will be used by Screen Reader to get node id information whenever
there is any action that needs to be perform at any local point on the
screen.

*******************************************************************
Apart from above changes, following are other changes in this cl which
were required for unit test:
* New Mock for Semantic Provider
* Refactoring of Semantic Action Listener.
* Unit Test for PerformHitTesting.

Testing:
  * Build: Passed
  * Unit Test: fx run_test a11y_tests (Passed)

MI4-2554 #comment

Change-Id: I40952cb7e758806a2a83d59c4a569541031bea26