[zircon] Split user-set affinity and kernel-set affinity.
Currently userspace and the kernel share the same affinity mask, with
either able to override the other. This isn't great, as often the kernel
will need to pin a thread to a particular CPU for reasons of
correctness.

This CL splits userspace and kernel affinities into two different masks.
Where possible, we try to respect both, but fall back to just respecting
the kernel's affinity mask if we can't.

BUG: ZX-3828
Change-Id: I9898388cd30aa9faa94a217c6adb2e5287265bbc