[zircon] Allow userspace to specify invalid CPU affinity masks.
Previously, we attempted to validate masks and fail profile creation if
the affinity mask was invalid. This isn't ideal: if/when CPUs go
offline, the mask may become invalid anyway. Instead, just let userspace
specify an arbitrary mask and feed it directly to the scheduler (which
no longer assumes the mask will be valid.)

Additionally, no longer promise that the mask fed into the kernel will
be the same as the mask returned. This avoids us having to stash the
entire bitmask just so we can hand it back to the user later.

BUG: ZX-3828
Change-Id: I5de767797afd5df49a0715f9d984d1b7687a72af