[garnet] Simple CLI tool for running command with given scheduler parameters.
This CL introduces a very simple CLI tool for running a given command
with the given scheduler parameters.

Currently, the only parameters that Zircon allows to be changed is
scheduler priority, but ZX-3828 will also add CPU affinity.

The tool is only very basic: it will adjust the first thread in a newly
launched job, but subsequent threads will just get the default system
scheduler priority.

BUG=ZX-3828

Change-Id: I36320fc6c83b1069f2b9358013baf16ccbee40e5