[kernel][x86] Enumerate swapgs leak bug (CVE-2019-1125)
SWAPGS is an x86 instruction to exchange the contents of the GS base
register and the IA32_KERNEL_GS_BASE MSR; it is used in the entry
and exit path of voluntary and involuntary kernel invocations, to
switch a user TLS pointer with a kernel per-cpu structure pointer.
Involuntary kernel entry paths conditionally execute SWAPGS, depending
on whether they interrupted user mode or kernel mode code.

Hostile user code could poison branch predictor state to cause the
SWAPGS to be skipped in speculative paths; it can also load the GS
base register via the WRGSBASE instruction with a hostile pointer,
to infer the contents of kernel memory via speculative gs-relative
kernel accesses.

Enumerate SWAPGS as present on all Intel x86-64 CPUs that are not
from the Atom (Bonnell/Silvermont/Airmont/Goldmont/...) family.

Change-Id: I019db592345229d33f0a6ee39efc8165f8669fe3