[debugger] Change Process attach flow.
Before the agent would notify the client immediatelly for any
pre-existing threads as "thread starting" events just after sending the
modules. This conflicts with the async nature of loading modules. This
will be specially true when the existent process has some thread with an
exception on it.

The new flow is that the client asks for threads when it's done loading
modules, so that there is no conflict between modules and thread
starting/exceptions.

TEST=Updated unit tests to reflect this.

Change-Id: I12503ad6004820a782147e84146229b524813fec