[DriverFramework][suspend] Fallback to reboot by default
When suspend task is stuck, currently we wait forever.
This is hindering infra workflows and so we should be rebooting
if the suspend timesout after 30sec. For debugging the driver suspend
issues, before kicking off the reboot, we dump the thread info of the
offending process.

ZX-4804:#Done

Test: Added automated test for suspend fail.
runtests -t devcoordinator-test. On Vim2, I manually failed xhci
driver suspend, and ensured that the debug info is printed, with the
callstacks and dso list of the offending devhost.

Change-Id: Ie9f5d10560302cc2612dbf9a7081653ac7347bde