[cloud_provider_firestore] fix polling CQ events
We're seeing interesting crashes where the gRPC completion queue returns
true in Next() w/o updating the pointer to the resulting tag. This
results in us casting a stale memory address to a callable and trying to
call that, with confusion and hilarity ensuing.

This patch fixes this by ensuring that we clear the result pointer each
time and only call it if it was actually set.

LE-721 #done
LE-820 #done

Change-Id: Icb99b257341f428fdd21f3f9e92d1fc46a867b41