[ledger] Add GetEntryId methods to the encryption service.
Add methods that generate an opaque entry id to identify an entry in a
diff exchanged with a cloud provider. This id is deterministically
generated for merge commits based on the commit ids of the parents,
entry name and the operation list; this is so that the same merge made
on different devices would use the same entry id.

Change-Id: I6714fef40852de2ff32506375fce540358cf1057