[ledger] track originating commit for tree nodes
With unordered storage, when we need to get an object from the network,
we have a different behavior if it's part of a tree node or of a value:
for values, we get them directly with GetObject, while for tree nodes,
fetching them from the cloud requires getting the whole diff for this
commit from the cloud and applying it.
Then, callers of GetObject need to indicate whether the object they are
requesting is part of a tree node or a value, and for tree nodes, what
commit it is a part of. This is done by extending the |location|
parameter with this information.

Change-Id: Ic1bf3070ac65616c7d063de11711806716da26cd