[netstack3] get|list interfaces returns addresses
Getting information for an interface and listing interfaces now returns
all IPv4 and IPv6 addresses configured on the interface.

NOTE: IPv6 tentative addresses are not returned because the FIDL
interface is just a list of addresses with no way to express that an
address is tentative.

TEST: modified `eventloop::integration_tests::test_get_interface_info`
and `eventloop::integration_tests::test_list_interfaces` to check that
the addresses in the interface properties returned are as expected.

Change-Id: I929356fcfe341b4922ff6213ae2812c7af9a61b1