[perftest] Disable printf()s of test progress, which skew test results
By default, the perftest library prints the name of each performance
test before running it, using printf().  Unfortunately, this causes
asynchronous background activity (e.g. of the network stack) which
makes the tests slower for some of the runs, producing outlier run
times.  This skews the test results and adds variation to the results,
producing wider confidence intervals in perfcompare's results, which
makes it harder to detect performance regressions and improvements.

Address this by adding and using a "--quiet" option that disables the
progress output.

The results from the perfcompare trybot for this change show:
  699 test cases in total
  574 test cases had no significant difference (no_sig_diff)
  124 test cases got faster
  1 test case got slower

This change might make debugging harder.  If it turns out to be
necessary, we could implement progress logging that prints output on a
time delay, to avoid the systematic interference that the current
logging has.

Issue: PT-245
Test: unit test: /boot/test/sys/perf-test
  + peridot perf trybot
  + fuchsia perfcompare trybot (see above)
Change-Id: I4b708dba57133826bc6cae514a0a589476dea378