[modular][config] Fix cloud provider xdr bug
FieldWithDefault() uses a bool to determine if it should read and write
from the data value (true) or the default value (false). We always
read the data value of cloud_provider into cloud_provider_str if it exists or
leave it as the default cloud provider value. Therefore, We should treat
cloud_provider_str as the default cloud provider when calling
FieldWithDefault().

TEST= fx run-test peridot_tests

Change-Id: I083788965c12043508d206b8ad926c5c4ae16ae7