[covargs] Support multiple inputs, use summmary as primary input
This changes the mode of operation for covargs. Previously we would use
the symbolizer output as the primary input and for each dump we would
try to find the corresponding test in the summary. This is problematic
because symbolizer output can contain dumps from executions of non-test
binaries which don't have corresponding summary entry. So instead we use
summary as the primary input and for each test we try to locate the
corresponding dump (which should be always present).  We also extend
processing to support multiple summaries and multiple symbolizer outputs
which is necessary for sharded executions.

Change-Id: I57323a5446d33b4ffb0d6e85153adeba7bceef20