[guest] use romfs for extras partition
This change adds a new mkromfs tool to generate a flat filesystem for
use in linux guests. This replaces the existing host-only mkextras
script, which required tools not present in the fuchsia toolchain. This
change also updates the virtmagma unit test to read from an ICD manifest
to locate the ICD file, rather than rely on a user-created symlink.

Test: ran virtmagma* tests from romfs disk
Change-Id: If45cd908d308b59c607bb45121fa6b9e70e39499