[docs] Clarify `fx update-rustc-crate-map` conditions.
While `fx update-rustc-crate-map` is listed after an `fx build` call, it
is not immediately clear that this command actually uses information
generated by the build.

I was making changes that only affected `Cargo.toml` and did not expect
`fx build` as a necesary requirement for `fx update-rustc-crate-map`.
`create_map.json` was not correctly updated and I was not sure where the
error is.

Ideally, we should fix `fx update-rustc-crate-map` to have proper
dependencies, and regenerate any out of date files.  But we can start
with, at least, documenting the dependency.

Change-Id: I547cae39d60fb09520b9865c6021c1af9e32620c